Got centroidal voronoi diagram working.
I rearranged everything so voronoi.js is only voronoi code, and it is now
more OOP style. Currently the code for determining voronoi regions is a
little crude, so I plan on fixing that. It only works for a small number
of generating points. If it gets too large there are errors that occur.

Removed a bunch of code that I'm not using. There are some things I'm
still figuring out so I'm unsure if they are needed.
	modified:   glUtils.js
	modified:   index.html
	modified:   voronoi.js